// src/app/core/models/Employee.ts
export interface Employee {
    id:              number;
    name:            string;
    lastname:        string;
    genderId:        number;
    maritalStatusId: number;
    birthdayDate:    Date;
    age:             number;
    dpi:             string;
    nit:             string;
    socialSecurity:  string;
    irtra:           string;
    passport:        string;
    address:         string;
    jobPositonID:    number;
    salary:          number;
    bonus:           number;
}